experiments/llama_rag_test/requirements.txt
# Llama実行用
llama-cpp-python==0.2.11

# RAGフレームワーク
langchain==0.0.335
langchain-community==0.0.1

# ベクトルDB
chromadb==0.4.18
faiss-cpu==1.8.0

# 埋め込みモデル
sentence-transformers==2.2.2

# ユーティリティ
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3